Safari-609.1.11.1/LayoutTests/http/tests/preload/link-header-on-subresource-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'http://127.0.0.1:8000/preload/resources/dummy-preloads-subresource.css.php' because non CSS MIME types are not allowed in strict mode.
PASS successfullyParsed is true

TEST COMPLETE
